web/requirements.txt
asgiref==3.5.2
certifi==2022.5.18.1
chardet==4.0.0
charset-normalizer==2.0.12
Django==2.2
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.0
environ==1.0
idna==3.3
#llvmlite==0.38.1
PyJWT==1.7.1
pytz==2022.1
python-dotenv
requests==2.27.1
sqlparse==0.4.2
urllib3==1.26.9
psycopg2-binary==2.9.1
aiogram
SpeechRecognition
requests
